Brings arcsde on trunk back to building state, and also forward-ports all changes from 2.4.x
git-svn-id: https://svn.osgeo.org/geotools/geotools/trunk@26536 e5c1c795-43da-0310-a71f-fac65c449510

+++ gt/modules/unsupported/arcsde/sde-dummy/src/main/java/com/esri/sde/sdk/client/SeColumnDefinition.java
@@ -25,7 +25,7 @@ public class SeColumnDefinition {
 	public int getType() { return 0; }
 	public int getSize() { return 0; }
 	public int getScale() { return 0; }
-	public int getRowIdType() { return 0; }
+	public short getRowIdType() { return 0; }
